Frontends/VirtualBox: get rid of vcc warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@35448 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIFrameBufferQImage.cpp
@@ -6,7 +6,7 @@
  */
 
 /*
- * Copyright (C) 2010 Oracle Corporation
+ * Copyright (C) 2010-2011 Oracle Corporation
  *
  * This file is part of VirtualBox Open Source Edition (OSE), as
  * available from http://www.virtualbox.org. This file is free software;
@@ -92,7 +92,7 @@ void UIFrameBufferQImage::paintEvent(QPaintEvent *pEvent)
 
     QPainter painter(m_pMachineView->viewport());
 
-    if (r.width() < m_width * 2 / 3)
+    if ((ulong)r.width() < m_width * 2 / 3)
     {
         /* This method is faster for narrow updates */
         m_PM = QPixmap::fromImage(pSourceImage->copy(r.x() + m_pMachineView->contentsX(),
